# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Taoist API
  description: |
    This is the API documentation for the Taoist platform, which provides services for both Laojun and Taishang domains.
    
    The Laojun domain focuses on plugin management, configuration management, and audit logging.
    The Taishang domain focuses on AI model management, vector collections, and task management.
  version: 1.0.0
  contact:
    name: Taoist API Support
    email: support@taoist.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.taoist.com
    description: Production server
...
paths:
...
  /api/taishang/models:
    get:
      tags:
        - Model
      summary: List models
      description: Retrieves a list of all models for the tenant
      operationId: listModels
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Model
      summary: Register model
      description: Registers a new AI model
      operationId: registerModel
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelCreateRequest'
      responses:
        '201':
          description: Model registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/taishang/models/{id}:
    get:
      tags:
        - Model
      summary: Get model
      description: Retrieves details of a specific model
      operationId: getModel
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Model
      summary: Update model
      description: Updates a specific model
      operationId: updateModel
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelUpdateRequest'
      responses:
        '200':
          description: Model updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Model
      summary: Delete model
      description: Deletes a specific model
      operationId: deleteModel
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Model deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Vector collection management
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    BaseResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
        traceId:
          type: string
          example: "abc123"
      required:
        - code
        - message
        - traceId
...
    MessageResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                message:
                  type: string
                  example: "Operation completed successfully"
              required:
                - message
          required:
            - data

    # Config schemas
...
    Model:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        tenant_id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174001"
        name:
          type: string
          example: "gpt-3.5-turbo"
        type:
          type: string
          example: "llm"
        description:
          type: string
          example: "OpenAI's GPT-3.5 Turbo model"
        status:
          type: string
          enum: ["active", "inactive"]
          example: "active"
        config:
          type: object
          example: {"api_key": "secret123", "temperature": 0.7}
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - tenant_id
        - name
        - type
        - status
    ModelResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Model'
          required:
            - data
    ModelListResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Model'
          required:
            - data
    ModelCreateRequest:
      type: object
      properties:
        name:
          type: string
          example: "gpt-3.5-turbo"
        type:
          type: string
          example: "llm"
        description:
          type: string
          example: "OpenAI's GPT-3.5 Turbo model"
        config:
          type: object
          example: {"api_key": "secret123", "temperature": 0.7}
      required:
        - name
        - type
    ModelUpdateRequest:
      type: object
      properties:
        name:
          type: string
          example: "gpt-3.5-turbo"
        type:
          type: string
          example: "llm"
        description:
          type: string
          example: "OpenAI's GPT-3.5 Turbo model"
        status:
          type: string
          enum: ["active", "inactive"]
          example: "active"
        config:
          type: object
          example: {"api_key": "secret123", "temperature": 0.7}

    # Vector collection schemas
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/BaseResponse'
              - type: object
                properties:
                  error:
                    type: string
                    example: "Invalid request parameters"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/BaseResponse'
              - type: object
                properties:
                  error:
                    type: string
                    example: "Invalid or expired token"
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/BaseResponse'
              - type: object
                properties:
                  error:
                    type: string
                    example: "Resource not found"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/BaseResponse'
              - type: object
                properties:
                  error:
                    type: string
                    example: "Internal server error"